CloudVerify/webapp-main/webapp/requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Werkzeug==3.0.4
bcrypt==4.0.1
requests==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.1
flask-bcrypt==1.0.1
email-validator==1.3.1
boto3==1.28.0
watchtower==2.0.1
statsd==3.3.0
sendgrid==6.9.7